requirements.txt
imutils==0.5.4
numpy==1.21.5
opencv_contrib_python==4.5.4.60
Pillow==9.1.0
selenium==3.141.0
tensorflow==2.8.0
tensorflow_hub==0.12.0
tqdm==4.62.3
